Add server_index for each remote_function_exec
When executing a command in a remote server, it might be useful to know
if the server is the first one or second or ...

This allows user to execute a function only on the first server.

Use case:
Given: I have the same cache server (memcached, redis, ...) for the
production server.
When: I want to deploy
Then I need to run a command once to flush the cache