Simplify git remote options
Configuration of repository-path and repository-url were over
complicated.
The goal was to allow the use of a bare repository as a source and use
that repository during the build process.
The only reason to use that configuration was when you hook the deploy
script to a git push. In other words, when you manage your own git
remote repository and want to deploy directly after a push is made on
your git repository.

This is an edge case that we no longer want to solve. People can rely on
some CI tools to perform the exact same action but not strictly when the
git repository receive the push.